+++ vector/src/main/java/im/vector/activity/JitsiCallActivity.java
@@ -27,15 +27,18 @@ import android.widget.FrameLayout;
 import com.facebook.react.modules.core.PermissionListener;
 
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.jitsi.meet.sdk.JitsiMeetActivityDelegate;
 import org.jitsi.meet.sdk.JitsiMeetActivityInterface;
 import org.jitsi.meet.sdk.JitsiMeetConferenceOptions;
+import org.jitsi.meet.sdk.JitsiMeetUserInfo;
 import org.jitsi.meet.sdk.JitsiMeetView;
 import org.jitsi.meet.sdk.JitsiMeetViewListener;
 import org.matrix.androidsdk.MXSession;
 import org.matrix.androidsdk.core.Log;
 import org.matrix.androidsdk.data.Room;
 
+import java.net.URL;
 import java.util.Map;
 
 import butterknife.BindView;
@@ -64,7 +67,7 @@ public class JitsiCallActivity extends VectorAppCompatActivity implements JitsiM
     /**
      * Base server URL
      */
-    private static final String JITSI_SERVER_URL = "https://jitsi.riot.im/";
+    public static final String JITSI_SERVER_URL = "https://jitsi.riot.im/";
 
     // the jitsi view
     private JitsiMeetView mJitsiView;
@@ -109,10 +112,7 @@ public class JitsiCallActivity extends VectorAppCompatActivity implements JitsiM
     @Override
     @SuppressLint("NewApi")
     public void initUiAndData() {
-        if (WidgetManagerProvider.INSTANCE.getWidgetManager(this) == null) {
-            finish();
-            return;
-        }
+
         // Waiting View
         setWaitingView(findViewById(R.id.jitsi_progress_layout));
 
@@ -136,7 +136,6 @@ public class JitsiCallActivity extends VectorAppCompatActivity implements JitsiM
             return;
         }
 
-
         mRoom = mSession.getDataHandler().getRoom(mWidget.getRoomId());
         if (null == mRoom) {
             Log.e(LOG_TAG, "## onCreate() : undefined room " + mWidget.getRoomId());
@@ -154,8 +153,22 @@ public class JitsiCallActivity extends VectorAppCompatActivity implements JitsiM
      */
     private void loadURL() {
         try {
+            JitsiMeetUserInfo userInfo = new JitsiMeetUserInfo();
+            userInfo.setDisplayName(mSession.getMyUser().displayname);
+            try {
+                String avatarUrl = mSession.getMyUser().avatar_url;
+                if (avatarUrl != null) {
+                    String downloadableUrl = mSession.getContentManager().getDownloadableUrl(avatarUrl, false);
+                    if (downloadableUrl != null) {
+                        userInfo.setAvatar(new URL(downloadableUrl));
+                    }
+                }
+            } catch (Exception e) {
+                //nop
+            }
             JitsiMeetConferenceOptions jitsiMeetConferenceOptions = new JitsiMeetConferenceOptions.Builder()
                     .setVideoMuted(!mIsVideoCall)
+                    .setUserInfo(userInfo)
                     // Configure the title of the screen
                     // TODO config.putString("callDisplayName", mRoom.getRoomDisplayName(this));
                     .setRoom(mCallUrl)
@@ -237,7 +250,7 @@ public class JitsiCallActivity extends VectorAppCompatActivity implements JitsiM
     protected void onStop() {
         super.onStop();
         JitsiMeetActivityDelegate.onHostPause(this);
-        WidgetsManager wm = WidgetManagerProvider.INSTANCE.getWidgetManager(this);
+        WidgetsManager wm = getWidgetManager();
         if (wm != null) {
             wm.removeListener(mWidgetListener);
         }
@@ -253,10 +266,18 @@ public class JitsiCallActivity extends VectorAppCompatActivity implements JitsiM
         super.onResume();
 
         JitsiMeetActivityDelegate.onHostResume(this);
-        WidgetsManager wm = WidgetManagerProvider.INSTANCE.getWidgetManager(this);
+        WidgetsManager wm = getWidgetManager();
         if (wm != null) {
             wm.addListener(mWidgetListener);
         }
+    }
+
+    @Nullable
+    private WidgetsManager getWidgetManager() {
+        if (mSession == null) return null;
+        WidgetManagerProvider widgetManagerProvider = Matrix.getInstance(this).getWidgetManagerProvider(mSession);
+        if (widgetManagerProvider == null) return null;
+        return widgetManagerProvider.getWidgetManager(this);
     }
 
     @Override