+++ vector/src/main/java/im/vector/fragments/VectorSettingsPreferencesFragment.kt
@@ -44,6 +44,7 @@ import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContextCompat
 import androidx.core.content.edit
 import androidx.core.view.isVisible
+import androidx.fragment.app.FragmentActivity
 import androidx.preference.*
 import com.bumptech.glide.Glide
 import com.google.android.material.textfield.TextInputEditText
@@ -60,11 +61,7 @@ import im.vector.dialogs.ExportKeysDialog
 import im.vector.extensions.getFingerprintHumanReadable
 import im.vector.extensions.showPassword
 import im.vector.extensions.withArgs
-import im.vector.fragments.troubleshoot.NotificationTroubleshootTestManager
-import im.vector.preference.ProgressBarPreference
-import im.vector.preference.UserAvatarPreference
-import im.vector.preference.VectorGroupPreference
-import im.vector.preference.VectorPreference
+import im.vector.preference.*
 import im.vector.settings.FontScale
 import im.vector.settings.VectorLocale
 import im.vector.ui.themes.ThemeUtils
@@ -74,6 +71,7 @@ import org.jetbrains.anko.toast
 import org.matrix.androidsdk.MXSession
 import org.matrix.androidsdk.call.MXCallsManager
 import org.matrix.androidsdk.core.BingRulesManager
+import org.matrix.androidsdk.core.JsonUtils
 import org.matrix.androidsdk.core.Log
 import org.matrix.androidsdk.core.ResourceUtils
 import org.matrix.androidsdk.core.callback.ApiCallback
@@ -88,10 +86,14 @@ import org.matrix.androidsdk.data.MyUser
 import org.matrix.androidsdk.data.Pusher
 import org.matrix.androidsdk.data.RoomMediaMessage
 import org.matrix.androidsdk.db.MXMediaCache
+import org.matrix.androidsdk.features.identityserver.IdentityServerManager
+import org.matrix.androidsdk.features.integrationmanager.IntegrationManager
 import org.matrix.androidsdk.listeners.MXEventListener
 import org.matrix.androidsdk.listeners.MXMediaUploadListener
+import org.matrix.androidsdk.rest.client.LoginRestClient
 import org.matrix.androidsdk.rest.model.bingrules.BingRule
 import org.matrix.androidsdk.rest.model.group.Group
+import org.matrix.androidsdk.rest.model.login.AuthParamsLoginPassword
 import org.matrix.androidsdk.rest.model.pid.ThirdPartyIdentifier
 import org.matrix.androidsdk.rest.model.pid.ThreePid
 import org.matrix.androidsdk.rest.model.sync.AccountDataElement
@@ -126,7 +128,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
         override fun onAccountDataUpdated(accountDataElement: AccountDataElement) {
             if (accountDataElement.type == AccountDataElement.ACCOUNT_DATA_TYPE_IDENTITY_SERVER) {
-                (findPreference(PreferencesManager.SETTINGS_IDENTITY_SERVER_PREFERENCE_KEY) as EditTextPreference).let {
+                (findPreference(PreferencesManager.SETTINGS_IDENTITY_SERVER_PREFERENCE_KEY) as VectorPreference).let {
                     updateIdentityServerPref()
                 }
             }
@@ -151,6 +153,12 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
     // current publicised group list
     private var mPublicisedGroups: MutableSet<String>? = null
+
+    private var integrationManagerManagerListener = object : IntegrationManager.IntegrationManagerManagerListener {
+        override fun onIntegrationManagerChange(managerConfig: IntegrationManager) {
+            refreshIntegrationManagerSettings()
+        }
+    }
 
     /* ==========================================================================================
      * Preferences
@@ -621,8 +629,33 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
         // identity server
         updateIdentityServerPref()
 
-        findPreference(PreferencesManager.SETTINGS_INTEGRATION_MANAGER_UI_URL)
-                .summary = PreferencesManager.getIntegrationManagerUiUrl(context)
+
+        (findPreference(PreferencesManager.SETTINGS_INTEGRATION_ALLOW) as? VectorSwitchPreference)?.let {
+            it.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, newValue ->
+
+                //Disable it while updating the state, will be re-enabled by the account data listener.
+                it.isEnabled = false
+                mSession.integrationManager.enableIntegrationManagerUsage(newValue as Boolean, object : ApiCallback<Void> {
+                    override fun onSuccess(info: Void?) {
+                        refreshIntegrationManagerSettings()
+                    }
+
+                    override fun onUnexpectedError(e: java.lang.Exception?) {
+                        refreshIntegrationManagerSettings()
+                    }
+
+                    override fun onNetworkError(e: java.lang.Exception?) {
+                        refreshIntegrationManagerSettings()
+                    }
+
+                    override fun onMatrixError(e: MatrixError?) {
+                        refreshIntegrationManagerSettings()
+                    }
+
+                })
+                true
+            }
+        }
 
         // Analytics
 
@@ -789,7 +822,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
                     } else {
                         MXCallsManager.defaultStunServerUri = null
                     }
-                    PreferencesManager.setUseDefaultTurnServer(activity,mUseDefaultStunPreference.isChecked)
+                    PreferencesManager.setUseDefaultTurnServer(activity, mUseDefaultStunPreference.isChecked)
                     false
                 }
             }
@@ -829,6 +862,29 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
             false
         }
+    }
+
+
+    private fun refreshIntegrationManagerSettings() {
+        val integrationAllowed = mSession.integrationManager.integrationAllowed
+        (findPreference(PreferencesManager.SETTINGS_INTEGRATION_ALLOW) as SwitchPreference).let {
+            val savedListener = it.onPreferenceChangeListener
+            it.onPreferenceChangeListener = null
+            it.isChecked = integrationAllowed
+            it.isEnabled = true
+            it.onPreferenceChangeListener = savedListener
+        }
+
+        findPreference(PreferencesManager.SETTINGS_INTEGRATION_MANAGER_UI_URL).let {
+            if (integrationAllowed) {
+                it.summary = Matrix.getWidgetManager(context)?.uiUrl ?: getString(R.string.none)
+                it.isVisible = true
+            } else {
+                it.isVisible = false
+            }
+        }
+
+
     }
 
     private fun refreshBackgroundSyncSection(appContext: Context?) {
@@ -896,7 +952,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
                                             // Even if using foreground service with foreground notif, it stops to work
                                             // in doze mode for certain devices :/
 
-                                            if ( !isIgnoringBatteryOptimizations(requireContext())) {
+                                            if (!isIgnoringBatteryOptimizations(requireContext())) {
                                                 requestDisablingBatteryOptimization(requireActivity(),
                                                         this@VectorSettingsPreferencesFragment,
                                                         REQUEST_BATTERY_OPTIMIZATION)
@@ -905,11 +961,11 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
                                             pushManager.setFdroidSyncModeOptimizedForRealTime();
                                         }
-                                        PreferencesManager.FDROID_BACKGROUND_SYNC_MODE_FOR_BATTERY -> {
+                                        PreferencesManager.FDROID_BACKGROUND_SYNC_MODE_FOR_BATTERY  -> {
 
                                             pushManager.setFdroidSyncModeOptimizedForBattery();
                                         }
-                                        PreferencesManager.FDROID_BACKGROUND_SYNC_MODE_DISABLED -> {
+                                        PreferencesManager.FDROID_BACKGROUND_SYNC_MODE_DISABLED     -> {
                                             pushManager.setFdroidSyncModeDisabled()
                                         }
                                     }
@@ -1031,6 +1087,9 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
             PreferenceManager.getDefaultSharedPreferences(context).registerOnSharedPreferenceChangeListener(this)
 
+            mSession.integrationManager.addListener(integrationManagerManagerListener)
+            refreshIntegrationManagerSettings()
+
             // refresh anything else
             refreshPreferences()
             refreshNotificationPrivacy()
@@ -1055,6 +1114,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
         if (mSession.isAlive) {
             mSession.dataHandler.removeListener(mEventsListener)
             Matrix.getInstance(context)?.removeNetworkEventListener(mNetworkListener)
+            mSession.integrationManager.removeListener(integrationManagerManagerListener)
         }
 
         PreferenceManager.getDefaultSharedPreferences(context).unregisterOnSharedPreferenceChangeListener(this)
@@ -1973,6 +2033,31 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
             addEmailBtn.order = order
             addEmailBtn.isVisible = true
+            addEmailBtn.isEnabled = false
+            //We need to check if the password flow is available
+            mSession.identityServerManager.checkAdd3pidInteractiveFlow(listOf(LoginRestClient.LOGIN_FLOW_TYPE_PASSWORD),
+                    object : ApiCallback<IdentityServerManager.SupportedFlowResult> {
+                        override fun onSuccess(info: IdentityServerManager.SupportedFlowResult) {
+                            addEmailBtn.isEnabled = info == IdentityServerManager.SupportedFlowResult.SUPPORTED
+                                    || info == IdentityServerManager.SupportedFlowResult.INTERACTIVE_AUTH_NOT_SUPPORTED
+                        }
+
+                        override fun onUnexpectedError(e: java.lang.Exception?) {
+                            //In doubt enable
+                            addEmailBtn.isEnabled = true
+                        }
+
+                        override fun onNetworkError(e: java.lang.Exception?) {
+                            //In doubt enable
+                            addEmailBtn.isEnabled = true
+                        }
+
+                        override fun onMatrixError(e: MatrixError?) {
+                            //In doubt enable
+                            addEmailBtn.isEnabled = true
+                        }
+
+                    })
         }
     }
 
@@ -1983,6 +2068,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
      * @param errorMessage the error message
      */
     private fun onCommonDone(errorMessage: String?) {
+        if (!isAdded) return
         activity?.runOnUiThread {
             if (!TextUtils.isEmpty(errorMessage) && errorMessage != null) {
                 VectorApp.getInstance().toast(errorMessage)
@@ -2017,7 +2103,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
         mSession.identityServerManager.startAddSessionForEmail(pid, null,
                 object : ApiCallback<ThreePid> {
                     override fun onSuccess(info: ThreePid) {
-                        activity?.runOnUiThread { showEmailValidationDialog(pid) }
+                        activity?.runOnUiThread { showEmailValidationDialog(pid, null) }
                     }
 
                     override fun onNetworkError(e: Exception) {
@@ -2043,50 +2129,102 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
      *
      * @param pid the used pid.
      */
-    private fun showEmailValidationDialog(pid: ThreePid) {
-        activity?.let {
-            AlertDialog.Builder(it)
-                    .setTitle(R.string.account_email_validation_title)
-                    .setMessage(R.string.account_email_validation_message)
-                    .setPositiveButton(R.string._continue) { _, _ ->
-                        mSession.identityServerManager.finalizeAddSessionForEmail(pid, object : ApiCallback<Void?> {
-                            override fun onSuccess(info: Void?) {
-                                it.runOnUiThread {
-                                    hideLoadingView()
-                                    mSession.myUser.refreshThirdPartyIdentifiers(object : SimpleApiCallback<Void?>(){
-                                        override fun onSuccess(info: Void?) {
-                                            refreshEmailsList()
-                                        }
-                                    })
-                                }
-                            }
-
-                            override fun onNetworkError(e: Exception) {
-                                onCommonDone(e.localizedMessage)
-                            }
-
-                            override fun onMatrixError(e: MatrixError) {
-                                if (TextUtils.equals(e.errcode, MatrixError.THREEPID_AUTH_FAILED)) {
-                                    it.runOnUiThread {
-                                        hideLoadingView()
-                                        it.toast(R.string.account_email_validation_error)
-                                    }
-                                } else {
-                                    onCommonDone(e.localizedMessage)
-                                }
-                            }
-
-                            override fun onUnexpectedError(e: Exception) {
-                                onCommonDone(e.localizedMessage)
-                            }
-                        })
-
-                    }
-                    .setNegativeButton(R.string.cancel) { _, _ ->
-                        hideLoadingView()
-                    }
-                    .show()
+    private fun showEmailValidationDialog(pid: ThreePid, password: String?) {
+        activity?.let { fragmentActivity ->
+            val auth = password?.let {
+                AuthParamsLoginPassword().apply {
+                    this.user = mSession.myUserId
+                    this.password = password
+                }
+            }
+            showEmailValidationDialog(fragmentActivity, pid, auth)
         }
+    }
+
+    private fun showEmailValidationDialog(fragmentActivity: FragmentActivity, pid: ThreePid, auth: AuthParamsLoginPassword?) {
+        AlertDialog.Builder(fragmentActivity)
+                .setTitle(R.string.account_email_validation_title)
+                .setMessage(R.string.account_email_validation_message)
+                .setPositiveButton(R.string._continue) { _, _ ->
+                    finalizeAdd(pid, auth, fragmentActivity)
+                }
+                .setNegativeButton(R.string.cancel) { _, _ ->
+                    hideLoadingView()
+                }
+                .show()
+    }
+
+    private fun finalizeAdd(pid: ThreePid, auth: AuthParamsLoginPassword?, fragmentActivity: FragmentActivity) {
+        mSession.identityServerManager.finalize3pidAddSession(pid, auth, object : ApiCallback<Void?> {
+            override fun onSuccess(info: Void?) {
+                if (!isAdded) return
+                fragmentActivity.runOnUiThread {
+                    hideLoadingView()
+                    mSession.myUser.refreshThirdPartyIdentifiers(object : SimpleApiCallback<Void?>() {
+                        override fun onSuccess(info: Void?) {
+                            refreshEmailsList()
+                        }
+                    })
+                }
+            }
+
+            override fun onNetworkError(e: Exception) {
+                onCommonDone(e.localizedMessage)
+            }
+
+            override fun onMatrixError(e: MatrixError) {
+                if (!isAdded) return
+                if (e.mStatus == 401 && e.mErrorBodyAsString.isNullOrBlank().not()) {
+                    val flow = JsonUtils.toRegistrationFlowResponse(e.mErrorBodyAsString)
+                    if (flow != null) {
+                        val supportsLoginPassword = flow.flows.any { it.stages == listOf(LoginRestClient.LOGIN_FLOW_TYPE_PASSWORD) }
+                        if (supportsLoginPassword) {
+                            //we prompt for it
+                            val invalidPassError = requireContext().getString(R.string.login_error_forbidden)
+                                    .takeIf { e.errcode == MatrixError.FORBIDDEN }
+                            DialogUtils.promptPassword(requireContext(),
+                                    invalidPassError,
+                                    auth?.let { auth.password },
+                                    { password ->
+                                        val authParams = AuthParamsLoginPassword().apply {
+                                            this.user = mSession.myUserId
+                                            this.password = password
+                                        }
+                                        finalizeAdd(pid, authParams, fragmentActivity)
+                                    },
+                                    {
+                                        onCommonDone(getString(R.string.settings_add_3pid_authentication_needed))
+                                    }
+                            )
+                        } else {
+                            //you can only do that on mobile
+                            AlertDialog.Builder(fragmentActivity)
+                                    .setMessage(R.string.settings_add_3pid_flow_not_supported)
+                                    .setPositiveButton(R.string._continue) { _, _ ->
+                                        onCommonDone(null)
+                                    }
+                                    .show()
+
+                        }
+
+                    } else {
+                        onCommonDone(e.localizedMessage)
+                    }
+                } else if (TextUtils.equals(e.errcode, MatrixError.THREEPID_AUTH_FAILED)) {
+                    fragmentActivity.runOnUiThread {
+                        fragmentActivity.toast(R.string.account_email_validation_error)
+                        //Re-display the popup so that not everything is lost
+                        showEmailValidationDialog(requireActivity(), pid, auth)
+                    }
+                } else {
+                    onCommonDone(e.localizedMessage)
+                }
+            }
+
+            override fun onUnexpectedError(e: Exception) {
+                onCommonDone(e.localizedMessage)
+            }
+        })
     }
 
     //==============================================================================================================
@@ -2219,6 +2357,31 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
             addPhoneBtn.order = order
             addPhoneBtn.isVisible = true
+            addPhoneBtn.isEnabled = false
+            //We need to check if the password flow is available
+            mSession.identityServerManager.checkAdd3pidInteractiveFlow(listOf(LoginRestClient.LOGIN_FLOW_TYPE_PASSWORD),
+                    object : ApiCallback<IdentityServerManager.SupportedFlowResult> {
+                        override fun onSuccess(info: IdentityServerManager.SupportedFlowResult) {
+                            addPhoneBtn.isEnabled = info == IdentityServerManager.SupportedFlowResult.SUPPORTED
+                                    || info == IdentityServerManager.SupportedFlowResult.INTERACTIVE_AUTH_NOT_SUPPORTED
+                        }
+
+                        override fun onUnexpectedError(e: java.lang.Exception?) {
+                            //In doubt enable
+                            addPhoneBtn.isEnabled = true
+                        }
+
+                        override fun onNetworkError(e: java.lang.Exception?) {
+                            //In doubt enable
+                            addPhoneBtn.isEnabled = true
+                        }
+
+                        override fun onMatrixError(e: MatrixError?) {
+                            //In doubt enable
+                            addPhoneBtn.isEnabled = true
+                        }
+
+                    })
         }
 
     }